Refactors member access and indexer resolution
Moves member access and indexer resolution logic from the parser to the type resolver.

This change centralizes these operations within the type resolver, promoting code reuse and improving maintainability.
It also removes duplicate code from the parser.
