+++ src/Hyperbee.XS/Core/TypeResolver.cs
@@ -2,7 +2,6 @@
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Runtime.CompilerServices;
-using System.Xml.Linq;
 using Hyperbee.Collections;
 
 namespace Hyperbee.XS.Core;
@@ -12,6 +11,8 @@ public interface ITypeResolver
     Type ResolveType( string typeName );
     MethodInfo ResolveMethod( Type type, string methodName, IReadOnlyList<Type> typeArgs, IReadOnlyList<Expression> args );
     MemberInfo ResolveMember( Type type, string memberName );
+    Expression ResolveIndexerExpression( Expression targetExpression, IReadOnlyList<Expression> indexes );
+    Expression ResolveMemberExpression( Expression targetExpression, string name, IReadOnlyList<Type> typeArgs, IReadOnlyList<Expression> args );
 }
 
 public class TypeResolver : ITypeResolver
@@ -165,6 +166,100 @@ public class TypeResolver : ITypeResolver
     {
         const BindingFlags BindingAttr = BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static;
         return type.GetMember( memberName, BindingAttr ).FirstOrDefault();
+    }
+
+    public virtual Expression ResolveIndexerExpression( Expression targetExpression, IReadOnlyList<Expression> indexes )
+    {
+        var indexers = targetExpression.Type.GetProperties()
+            .Where( p => p.GetIndexParameters().Length == indexes.Count )
+            .ToArray();
+
+        if ( indexers.Length == 0 )
+            return Expression.ArrayAccess( targetExpression, indexes );
+
+        // Find the best match based on parameter types
+        var indexer = indexers.FirstOrDefault( p =>
+            p.GetIndexParameters()
+                .Select( param => param.ParameterType )
+                .SequenceEqual( indexes.Select( i => i.Type ) ) );
+
+        if ( indexer == null )
+        {
+            throw new InvalidOperationException(
+                $"No matching indexer found on type '{targetExpression.Type}' with parameter types: " +
+                $"{string.Join( ", ", indexes.Select( i => i.Type.Name ) )}." );
+        }
+
+        return Expression.Property( targetExpression, indexer, indexes.ToArray() );
+    }
+
+    public virtual Expression ResolveMemberExpression( Expression targetExpression, string name, IReadOnlyList<Type> typeArgs, IReadOnlyList<Expression> args )
+    {
+        var type = TypeOf( targetExpression );
+
+        // method
+
+        if ( args != null )
+        {
+            var method = ResolveMethod( type, name, typeArgs, args );
+
+            if ( method == null )
+                throw new InvalidOperationException( $"Method '{name}' not found on type '{type}'." );
+
+            var arguments = GetArgumentsWithDefaults( method, targetExpression, args );
+
+            return method.IsStatic
+                ? Expression.Call( method, arguments )
+                : Expression.Call( targetExpression, method, arguments );
+        }
+
+        // property or field
+
+        var member = ResolveMember( type, name );
+
+        if ( member == null )
+            throw new InvalidOperationException( $"Member '{name}' not found on type '{type}'." );
+
+        return member switch
+        {
+            PropertyInfo property => Expression.Property( targetExpression, property ),
+            FieldInfo field => Expression.Field( targetExpression, field ),
+            _ => throw new InvalidOperationException( $"Member '{name}' is not a property or field." )
+        };
+
+        static IReadOnlyList<Expression> GetArgumentsWithDefaults( MethodInfo method, Expression targetExpression, IReadOnlyList<Expression> providedArgs )
+        {
+            var parameters = method.GetParameters();
+            var isExtension = method.IsDefined( typeof( ExtensionAttribute ), false );
+
+            var providedOffset = isExtension ? 1 : 0;
+            var providedCount = providedArgs.Count;
+            var totalParameters = parameters.Length;
+
+            if ( providedCount == totalParameters )
+                return providedArgs;
+
+            var methodArgs = new Expression[totalParameters];
+
+            // add provided arguments
+            if ( isExtension )
+                methodArgs[0] = targetExpression;
+
+            for ( var i = 0; i < providedCount; i++ )
+            {
+                methodArgs[i + providedOffset] = providedArgs[i];
+            }
+
+            // add missing optional parameters
+            for ( var i = providedCount + providedOffset; i < totalParameters; i++ )
+            {
+                methodArgs[i] = parameters[i].HasDefaultValue
+                    ? Expression.Constant( parameters[i].DefaultValue, parameters[i].ParameterType )
+                    : throw new ArgumentException( $"Missing required parameter: {parameters[i].Name}" );
+            }
+
+            return methodArgs;
+        }
     }
 
     public void RegisterExtensionMethods( IEnumerable<Assembly> assemblies )
@@ -400,5 +495,18 @@ public class TypeResolver : ITypeResolver
 
         return p1.Length.CompareTo( p2.Length );
     }
+
+    [MethodImpl( MethodImplOptions.AggressiveInlining )]
+    private static Type TypeOf( Expression expression )
+    {
+        ArgumentNullException.ThrowIfNull( expression, nameof( expression ) );
+
+        return expression switch
+        {
+            ConstantExpression ce => ce.Value as Type ?? ce.Type,
+            _ => expression.Type
+        };
+    }
+
 }
